Restructuring of Apparatus fontium
I have added some general introductory matter, hopefully giving a better
overview of the structure and made some things a bit more explicit.
I have also pulled in the suggestions from issue #54.

I don't consider this to be finished yet, so any changes, suggestions or
thoughts are very welcome.

I especially wonder whether we might want to make a small write up of
how to integrate external prosopographical lists into the processing.
And not least the creation of canonical references (which I still
consider to be quite a challenge, as you know).