Reordering of list and rephrase cf. issue #59
The reordering of the list makes the presentation of no. 7 work
again (the long bullet point and the example broke the list count).

I have rephrased the description a bit to be a bit more explicit about
the use of empty lemmas (their purpose) in relation to the established
text. This could close issue #59.