Add whitespace after ending quotation mark
I suggest this because if you convert this to RNC (which should be
possible, as it is required in some contexts), you will end up with the
illegal formula """... e.g. "lbp-critical-1.0.0""""

This contains four closing quotations marks which is illegal.

With a whitespace here we end up with
""" ... e.g. "lbp-critical-1.0.0" """
And that is okay.